+++ virtualizarr/manifests/store.py
@@ -2,7 +2,8 @@ from __future__ import annotations
 
 import pickle
 from collections.abc import Iterable
-from typing import TYPE_CHECKING, Any
+from typing import TYPE_CHECKING, Any, Mapping
+from urllib.parse import urlparse
 
 from zarr.abc.store import (
     ByteRequest,
@@ -14,6 +15,7 @@ from zarr.abc.store import (
 from zarr.core.buffer import Buffer
 from zarr.core.buffer.core import BufferPrototype
 
+from virtualizarr.manifests.array import ManifestArray
 from virtualizarr.manifests.group import ManifestGroup
 
 if TYPE_CHECKING:
@@ -25,6 +27,7 @@ if TYPE_CHECKING:
 
 
 __all__ = ["ManifestStore"]
+
 
 _ALLOWED_EXCEPTIONS: tuple[type[Exception], ...] = (
     FileNotFoundError,
@@ -38,13 +41,14 @@ from typing import TYPE_CHECKING, Any, TypeAlias
 
 from zarr.core.buffer import default_buffer_prototype
 
-from virtualizarr.manifests.group import ManifestArrayVariableMapping
-from virtualizarr.vendor.zarr.core.metadata import dict_to_buffer
+from virtualizarr.vendor.zarr.metadata import dict_to_buffer
 
 if TYPE_CHECKING:
-    from obstore.store import ObjectStore
+    from obstore.store import ObjectStore  # type: ignore[import-not-found]
 
     StoreDict: TypeAlias = dict[str, ObjectStore]
+
+    import xarray as xr
 
 
 @dataclass
@@ -58,20 +62,20 @@ class StoreRequest:
 
 
 async def list_dir_from_manifest_arrays(
-    manifest_arrays: ManifestArrayVariableMapping, prefix: str
+    arrays: Mapping[str, ManifestArray], prefix: str
 ) -> AsyncGenerator[str]:
     """Create the expected results for Zarr's `store.list_dir()` from an Xarray DataArrray or Dataset
 
     Parameters
     ----------
-    manifest_arrays : ManifestArrayVariableMapping
+    arrays : Mapping[str, ManifestArrays]
     prefix : str
-
 
     Returns
     -------
     AsyncIterator[str]
     """
+    # TODO shouldn't this just accept a ManifestGroup instead?
     # Start with expected group level metadata
     raise NotImplementedError
 
@@ -98,25 +102,20 @@ def get_zarr_metadata(manifest_group: ManifestGroup, key: str) -> Buffer:
     # If requesting the root metadata, return the standard group metadata with additional dataset specific attributes
 
     if key == "zarr.json":
-        metadata = manifest_group._metadata.to_dict()
+        metadata = manifest_group.metadata.to_dict()
         return dict_to_buffer(metadata, prototype=default_buffer_prototype())
     else:
         var, _ = key.split("/")
-        metadata = manifest_group._manifest_arrays[var].metadata.to_dict()
+        metadata = manifest_group.arrays[var].metadata.to_dict()
         return dict_to_buffer(metadata, prototype=default_buffer_prototype())
 
 
-def parse_manifest_index(
-    key: str, chunk_key_encoding: str = "."
-) -> tuple[str, tuple[int, ...]]:
+def parse_manifest_index(key: str, chunk_key_encoding: str = ".") -> tuple[int, ...]:
     """
     Splits `key` provided to a zarr store into the variable indicated
     by the first part and the chunk index from the 3rd through last parts,
     which can be used to index into the ndarrays containing paths, offsets,
     and lengths in ManifestArrays.
-
-    Currently only works for 1d+ arrays with a tree depth of one from the
-    root Zarr group.
 
     Parameters
     ----------
@@ -125,17 +124,17 @@ def parse_manifest_index(
 
     Returns
     -------
-    ManifestIndex
+    tuple containing chunk indexes
     """
-    parts = key.split("/")
-    var = parts[0]
-    # Assume "c" is the second part
-    # TODO: Handle scalar array case with "c" holds the data
-    if chunk_key_encoding == "/":
-        indexes = tuple(int(ind) for ind in parts[2:])
-    else:
-        indexes = tuple(int(ind) for ind in parts[2].split(chunk_key_encoding))
-    return var, indexes
+    if key.endswith("c"):
+        # Scalar arrays hold the data in the "c" key
+        raise NotImplementedError(
+            "Scalar arrays are not yet supported by ManifestStore"
+        )
+    parts = key.split(
+        "c/"
+    )  # TODO: Open an issue upstream about the Zarr spec indicating this should be f"c{chunk_key_encoding}" rather than always "c/"
+    return tuple(int(ind) for ind in parts[1].split(chunk_key_encoding))
 
 
 def find_matching_store(stores: StoreDict, request_key: str) -> StoreRequest:
@@ -160,7 +159,8 @@ def find_matching_store(stores: StoreDict, request_key: str) -> StoreRequest:
     # Check each key to see if it's a prefix of the uri_string
     for key in sorted_keys:
         if request_key.startswith(key):
-            return StoreRequest(store=stores[key], key=request_key[len(key) :])
+            parsed_key = urlparse(request_key)
+            return StoreRequest(store=stores[key], key=parsed_key.path)
     # if no match is found, raise an error
     raise ValueError(
         f"Expected the one of stores.keys() to match the data prefix, got {stores.keys()} and {request_key}"
@@ -168,16 +168,17 @@ def find_matching_store(stores: StoreDict, request_key: str) -> StoreRequest:
 
 
 class ManifestStore(Store):
-    """A read-only Zarr store that uses obstore to access data on AWS, GCP, Azure. The requests
+    """
+    A read-only Zarr store that uses obstore to access data on AWS, GCP, Azure. The requests
     from the Zarr API are redirected using the :class:`virtualizarr.manifests.ManifestGroup` containing
     multiple :class:`virtualizarr.manifests.ManifestArray`,
     allowing for virtually interfacing with underlying data in other file format.
 
-
     Parameters
     ----------
-    manifest_group : ManifestGroup
-        Manifest Group containing Group metadata and mapping variable names to ManifestArrays
+    group : ManifestGroup
+        Root group of the store.
+        Contains group metadata, ManifestArrays, and any subgroups.
     stores : dict[prefix, :class:`obstore.store.ObjectStore`]
         A mapping of url prefixes to obstore Store instances set up with the proper credentials.
 
@@ -194,7 +195,7 @@ class ManifestStore(Store):
     Modified from https://github.com/zarr-developers/zarr-python/pull/1661
     """
 
-    _manifest_group: ManifestGroup
+    _group: ManifestGroup
     _stores: StoreDict
 
     def __eq__(self, value: object):
@@ -202,7 +203,7 @@ class ManifestStore(Store):
 
     def __init__(
         self,
-        manifest_group: ManifestGroup,
+        group: ManifestGroup,
         *,
         stores: StoreDict,  # TODO: Consider using a sequence of tuples rather than a dict (see https://github.com/zarr-developers/VirtualiZarr/pull/490#discussion_r2010717898).
     ) -> None:
@@ -221,14 +222,17 @@ class ManifestStore(Store):
         for store in stores.values():
             if not store.__class__.__module__.startswith("obstore"):
                 raise TypeError(f"expected ObjectStore class, got {store!r}")
+
         # TODO: Don't allow stores with prefix
-        # TODO: Type check the manifest arrays
+        if not isinstance(group, ManifestGroup):
+            raise TypeError
+
         super().__init__(read_only=True)
         self._stores = stores
-        self._manifest_group = manifest_group
+        self._group = group
 
     def __str__(self) -> str:
-        return f"ManifestStore({self._manifest_group}, {self._stores})"
+        return f"ManifestStore(group={self._group}, stores={self._stores})"
 
     def __getstate__(self) -> dict[Any, Any]:
         state = self.__dict__.copy()
@@ -255,14 +259,17 @@ class ManifestStore(Store):
         import obstore as obs
 
         if key.endswith("zarr.json"):
-            return get_zarr_metadata(self._manifest_group, key)
-        var, chunk_key = parse_manifest_index(key)
-        marr = self._manifest_group._manifest_arrays[var]
-        manifest = marr._manifest
+            return get_zarr_metadata(self._group, key)
+        var = key.split("/")[0]
+        marr = self._group.arrays[var]
+        manifest = marr.manifest
 
-        path = manifest._paths[*chunk_key]
-        offset = manifest._offsets[*chunk_key]
-        length = manifest._lengths[*chunk_key]
+        chunk_indexes = parse_manifest_index(
+            key, marr.metadata.chunk_key_encoding.separator
+        )
+        path = manifest._paths[*chunk_indexes]
+        offset = manifest._offsets[*chunk_indexes]
+        length = manifest._lengths[*chunk_indexes]
         # Get the  configured object store instance that matches the path
         store_request = find_matching_store(stores=self._stores, request_key=path)
         # Transform the input byte range to account for the chunk location in the file
@@ -343,8 +350,42 @@ class ManifestStore(Store):
     async def list_dir(self, prefix: str) -> AsyncGenerator[str, None]:
         # docstring inherited
         yield "zarr.json"
-        for k in self._manifest_group._manifest_arrays.keys():
+        for k in self._group.arrays.keys():
             yield k
+
+    def to_virtual_dataset(
+        self,
+        group="",
+        loadable_variables: Iterable[str] | None = None,
+        decode_times: bool | None = None,
+        indexes: Mapping[str, xr.Index] | None = None,
+    ) -> "xr.Dataset":
+        """
+        Create a "virtual" xarray dataset containing the contents of one zarr group.
+
+        Will ignore the contents of any other groups in the store.
+
+        Requires xarray.
+
+        Parameters
+        ----------
+        group : str
+        loadable_variables : Iterable[str], optional
+
+        Returns
+        -------
+        vds : xarray.Dataset
+        """
+
+        from virtualizarr.xarray import construct_virtual_dataset
+
+        return construct_virtual_dataset(
+            manifest_store=self,
+            group=group,
+            loadable_variables=loadable_variables,
+            indexes=indexes,
+            decode_times=decode_times,
+        )
 
 
 def _transform_byte_range(