+++ virtualizarr/readers/fits.py
@@ -6,12 +6,12 @@ from xarray import Dataset, Index
 from virtualizarr.readers.api import (
     VirtualBackend,
 )
-from virtualizarr.readers.common import construct_fully_virtual_dataset
 from virtualizarr.translators.kerchunk import (
     extract_group,
     virtual_vars_and_metadata_from_kerchunk_refs,
 )
 from virtualizarr.types.kerchunk import KerchunkStoreRefs
+from virtualizarr.xarray import construct_fully_virtual_dataset
 
 
 class FITSVirtualBackend(VirtualBackend):