+++ virtualizarr/tests/test_backend.py
@@ -1,4 +1,6 @@
+import functools
 from collections.abc import Mapping
+from concurrent.futures import ThreadPoolExecutor
 from pathlib import Path
 from unittest.mock import patch
 
@@ -9,7 +11,7 @@ import xarray.testing as xrt
 from xarray import Dataset, open_dataset
 from xarray.core.indexes import Index
 
-from virtualizarr import open_virtual_dataset
+from virtualizarr import open_virtual_dataset, open_virtual_mfdataset
 from virtualizarr.backend import (
     FileType,
     VirtualBackend,
@@ -21,8 +23,10 @@ from virtualizarr.readers.hdf import HDFVirtualBackend
 from virtualizarr.tests import (
     has_astropy,
     parametrize_over_hdf_backends,
+    requires_dask,
     requires_hdf5plugin,
     requires_imagecodecs,
+    requires_lithops,
     requires_network,
     requires_s3fs,
     requires_scipy,
@@ -405,7 +409,7 @@ class TestLoadVirtualDataset:
             xr.open_dataset(netcdf4_file, decode_times=True) as ds,
         ):
             assert set(vds.variables) == set(ds.variables)
-            assert set(vds.coords) == set(vds.coords)
+            assert set(vds.coords) == set(ds.coords)
 
             virtual_variables = {
                 name: var
@@ -420,7 +424,7 @@ class TestLoadVirtualDataset:
 
             assert set(actual_loadable_variables) == set(expected_loadable_variables)
 
-            for name, var in virtual_variables.items():
+            for var in virtual_variables.values():
                 assert isinstance(var.data, ManifestArray)
 
             for name, var in ds.variables.items():
@@ -502,3 +506,88 @@ class TestLoadVirtualDataset:
         with open_virtual_dataset(hdf5_scalar, backend=hdf_backend) as vds:
             assert vds.scalar.dims == ()
             assert vds.scalar.attrs == {"scalar": "true"}
+
+
+preprocess_func = functools.partial(
+    xr.Dataset.rename_vars,
+    air="nair",
+)
+
+
+@requires_hdf5plugin
+@requires_imagecodecs
+@parametrize_over_hdf_backends
+class TestOpenVirtualMFDataset:
+    @pytest.mark.parametrize("invalid_parallel_kwarg", ["ray", Dataset])
+    def test_invalid_parallel_kwarg(
+        self, netcdf4_files_factory, invalid_parallel_kwarg, hdf_backend
+    ):
+        filepath1, filepath2 = netcdf4_files_factory()
+
+        with pytest.raises(ValueError, match="Unrecognized argument"):
+            open_virtual_mfdataset(
+                [filepath1, filepath2],
+                combine="nested",
+                concat_dim="time",
+                backend=hdf_backend,
+                parallel=invalid_parallel_kwarg,
+            )
+
+    @pytest.mark.parametrize(
+        "parallel",
+        [
+            False,
+            ThreadPoolExecutor,
+            pytest.param("dask", marks=requires_dask),
+            pytest.param("lithops", marks=requires_lithops),
+        ],
+    )
+    @pytest.mark.parametrize(
+        "preprocess",
+        [
+            None,
+            preprocess_func,
+        ],
+    )
+    def test_parallel_open(
+        self, netcdf4_files_factory, hdf_backend, parallel, preprocess
+    ):
+        filepath1, filepath2 = netcdf4_files_factory()
+        vds1 = open_virtual_dataset(filepath1, backend=hdf_backend)
+        vds2 = open_virtual_dataset(filepath2, backend=hdf_backend)
+
+        expected_vds = xr.concat([vds1, vds2], dim="time")
+        if preprocess:
+            expected_vds = preprocess_func(expected_vds)
+
+        # test combine nested, which doesn't use in-memory indexes
+        combined_vds = open_virtual_mfdataset(
+            [filepath1, filepath2],
+            combine="nested",
+            concat_dim="time",
+            backend=hdf_backend,
+            parallel=parallel,
+            preprocess=preprocess,
+        )
+        xrt.assert_identical(combined_vds, expected_vds)
+
+        # test combine by coords using in-memory indexes
+        combined_vds = open_virtual_mfdataset(
+            [filepath1, filepath2],
+            combine="by_coords",
+            backend=hdf_backend,
+            parallel=parallel,
+            preprocess=preprocess,
+        )
+        xrt.assert_identical(combined_vds, expected_vds)
+
+        # test combine by coords again using in-memory indexes but for a glob
+        file_glob = Path(filepath1).parent.glob("air*.nc")
+        combined_vds = open_virtual_mfdataset(
+            file_glob,
+            combine="by_coords",
+            backend=hdf_backend,
+            parallel=parallel,
+            preprocess=preprocess,
+        )
+        xrt.assert_identical(combined_vds, expected_vds)